Move facility map rendering outside map comp
